Drop auth module
