+++ up42/__init__.py
@@ -21,7 +21,6 @@ from typing import Callable, Type, Union, cast
 
 # pylint: disable=only-importing-modules-is-allowed
 from up42.asset import Asset
-from up42.auth import Auth
 from up42.base import authenticate, get_credits_balance
 from up42.catalog import Catalog
 from up42.glossary import CollectionSorting, CollectionType, ProductGlossary
@@ -48,7 +47,6 @@ __all__ = [
     ).__name__
     for obj in [
         Asset,
-        Auth,
         Catalog,
         Order,
         Storage,
